Add Windows test, added all helpers functions and some tests

+++ deployability/modules/testing/tests/helpers/generic.py
@@ -12,7 +12,7 @@ import time
 import yaml
 
 from pathlib import Path
-from .constants import WAZUH_CONTROL, CLIENT_KEYS
+from .constants import WAZUH_CONTROL, CLIENT_KEYS, WINDOWS_CLIENT_KEYS, WINDOWS_VERSION, WINDOWS_REVISION
 from .executor import Executor
 from modules.generic.logger import logger
 from .utils import Utils
@@ -32,7 +32,11 @@ class HostInformation:
         Returns:
             bool: True or False
         """
-        return 'true' in Executor.execute_command(inventory_path, f'test -d {dir_path} && echo "true" || echo "false"')
+        os_type = HostInformation.get_os_type(inventory_path)
+        if os_type == 'linux':
+            return 'true' in Executor.execute_commands(inventory_path, f'test -d {dir_path} && echo "true" || echo "false"')
+        elif os_type == 'windows':
+            return 'true' in Executor.execute_commands(inventory_path, f'Test-Path -Path "{dir_path}"')
 
 
     @staticmethod
@@ -47,7 +51,7 @@ class HostInformation:
         Returns:
             bool: True or False
         """
-        return 'true' in Executor.execute_command(inventory_path, f'test -f {file_path} && echo "true" || echo "false"')
+        return 'true' in Executor.execute_commands(inventory_path, f'Test-Path -Path "{file_path}"')
 
 
     @staticmethod
@@ -61,8 +65,19 @@ class HostInformation:
         Returns:
             str: type of host (windows, linux, macos)
         """
-        system = Executor.execute_command(inventory_path, 'uname')
-        return system.lower()
+        try:
+            with open(inventory_path.replace('inventory', 'track'), 'r') as file:
+                data = yaml.safe_load(file)
+            if 'platform' in data:
+                return data['platform']
+            else:
+                raise KeyError("The 'platform' key was not found in the YAML file.")
+        except FileNotFoundError:
+            logger.error(f"The YAML file '{inventory_path}' was not found.")
+        except yaml.YAMLError as e:
+            logger.error(f"Error while loading the YAML file: {e}")
+        except Exception as e:
+            logger.error(f"An unexpected error occurred: {e}")
 
 
     @staticmethod
@@ -74,9 +89,21 @@ class HostInformation:
             inventory_path: host's inventory path
 
         Returns:
-            str: architecture (aarch64, x86_64, intel, apple)
+            str: architecture (amd64, arm64, intel, apple)
         """
-        return Executor.execute_command(inventory_path, 'uname -m')
+        try:
+            with open(inventory_path.replace('inventory', 'track'), 'r') as file:
+                data = yaml.safe_load(file)
+            if 'platform' in data:
+                return data['arch']
+            else:
+                raise KeyError("The 'platform' key was not found in the YAML file.")
+        except FileNotFoundError:
+            logger.error(f"The YAML file '{inventory_path}' was not found.")
+        except yaml.YAMLError as e:
+            logger.error(f"Error while loading the YAML file: {e}")
+        except Exception as e:
+            logger.error(f"An unexpected error occurred: {e}")
 
 
     @staticmethod
@@ -115,9 +142,9 @@ class HostInformation:
             str: linux os name (debian, ubuntu, opensuse, amazon, centos, redhat)
         """
         if 'manager' in inventory_path:
-            os_name = re.search(r'/manager-linux-([^-]+)-', inventory_path).group(1)
+            os_name = re.search(r'/manager-[^-]+-([^-]+)-', inventory_path).group(1)
         elif 'agent' in inventory_path:
-            os_name = re.search(r'/agent-linux-([^-]+)-', inventory_path).group(1)
+            os_name = re.search(r'/agent-[^-]+-([^-]+)-', inventory_path).group(1)
 
         return os_name
 
@@ -133,9 +160,9 @@ class HostInformation:
             tuple: linux os name and version (e.g., ('ubuntu', '22.04'))
         """
         if 'manager' in inventory_path:
-            match = re.search(r'/manager-linux-([^-]+)-([^-]+)-', inventory_path)
+            match = re.search(r'/manager-[^-]+-([^-]+)-([^-]+)-', inventory_path)
         elif 'agent' in inventory_path:
-            match = re.search(r'/agent-linux-([^-]+)-([^-]+)-', inventory_path)
+            match = re.search(r'/agent-[^-]+-([^-]+)-([^-]+)-', inventory_path)
         if match:
             os_name = match.group(1)
             version = match.group(2)
@@ -154,8 +181,12 @@ class HostInformation:
         Returns:
             str: current directory
         """
+        os_type = HostInformation.get_os_type(inventory_path)
 
-        return Executor.execute_command(inventory_path, 'pwd').replace("\n","")
+        if os_type == 'linux':
+            return Executor.execute_commands(inventory_path, 'pwd').replace("\n","")
+        elif os_type == 'windows':
+            return Executor.execute_commands(inventory_path, '(Get-Location).Path').replace("\n","")
 
     @staticmethod
     def get_internal_ip_from_aws_dns(dns_name):
@@ -188,7 +219,14 @@ class HostInformation:
             list: List of dictionaries with the client keys.
         """
         clients = []
-        client_key = Executor.execute_command(inventory_path, f'cat {CLIENT_KEYS}')
+
+        os_type = HostInformation.get_os_type(inventory_path)
+
+        if os_type == 'linux':
+            client_key = Executor.execute_commands(inventory_path, f'cat {CLIENT_KEYS}')
+        elif os_type == 'windows':
+            client_key = Executor.execute_commands(inventory_path, f'Get-Content {WINDOWS_CLIENT_KEYS}')
+
         lines = client_key.split('\n')[:-1]
         for line in lines:
             _id, name, address, password = line.strip().split()
@@ -227,14 +265,21 @@ class HostConfiguration:
             inventory_path: host's inventory path
 
         """
-        commands = ["sudo systemctl stop firewalld", "sudo systemctl disable firewalld"]
-        if GeneralComponentActions.isComponentActive(inventory_path, 'firewalld'):
+        commands = []
+
+        os_type = HostInformation.get_os_type(inventory_path)
+
+        if os_type == 'linux':
+            commands = ["sudo systemctl stop firewalld", "sudo systemctl disable firewalld"]
+            if GeneralComponentActions.isComponentActive(inventory_path, 'firewalld'):
+                Executor.execute_commands(inventory_path, commands)
+
+                logger.info(f'Firewall disabled on {HostInformation.get_os_name_and_version_from_inventory(inventory_path)}')
+            else:
+                logger.info(f'No Firewall to disable on {HostInformation.get_os_name_and_version_from_inventory(inventory_path)}')
+        elif os_type == 'windows':
+            commands = ["Set-NetFirewallProfile -Profile Domain,Public,Private -Enabled False"]
             Executor.execute_commands(inventory_path, commands)
-
-            logger.info(f'Firewall disabled on {HostInformation.get_os_name_and_version_from_inventory(inventory_path)}')
-        else:
-            logger.info(f'No Firewall to disable on {HostInformation.get_os_name_and_version_from_inventory(inventory_path)}')
-
 
 
     def _extract_hosts(paths, is_aws):
@@ -354,7 +399,7 @@ class HostConfiguration:
 
         # Allowing handling permissions
         if file_name == 'wazuh-install-files.tar':
-            Executor.execute_command(from_inventory_path, f'chmod +rw {file_name}')
+            Executor.execute_commands(from_inventory_path, f'chmod +rw {file_name}')
             logger.info('File permissions modified to be handled')
 
         # SCP
@@ -371,8 +416,8 @@ class HostConfiguration:
 
         # Restoring permissions
         if file_name == 'wazuh-install-files.tar':
-            Executor.execute_command(from_inventory_path, f'chmod 600 {file_name}')
-            Executor.execute_command(to_inventory_path, f'chmod 600 {file_name}')
+            Executor.execute_commands(from_inventory_path, f'chmod 600 {file_name}')
+            Executor.execute_commands(to_inventory_path, f'chmod 600 {file_name}')
             logger.info('File permissions were restablished')
 
         # Deleting file from localhost
@@ -459,7 +504,7 @@ class CheckFiles:
         """
         if 'linux' in os_type or 'macos' in os_type:
             command = f'sudo find {directory} -type f -exec sha256sum {{}} + {filter}'
-            result = Executor.execute_command(inventory_path, command)
+            result = Executor.execute_commands(inventory_path, command)
 
         elif 'windows' in os_type:
             command = 'dir /a-d /b /s | findstr /v /c:"\\.$" /c:"\\..$"| find /c ":"'
@@ -533,8 +578,13 @@ class GeneralComponentActions:
             str: Role status
         """
         logger.info(f'Getting status of {HostInformation.get_os_name_and_version_from_inventory(inventory_path)}')
+        os_type = HostInformation.get_os_type(inventory_path)
 
-        return Executor.execute_command(inventory_path, f'systemctl status {host_role}')
+        if os_type == 'linux':
+            return Executor.execute_commands(inventory_path, f'systemctl status {host_role}')
+        elif os_type == 'windows':
+            command = "Get-Service -Name 'Wazuh' | Format-Table -HideTableHeaders Status"
+            return Executor.execute_commands(inventory_path, command.replace("\n", ""))
 
 
     @staticmethod
@@ -548,7 +598,12 @@ class GeneralComponentActions:
         """
 
         logger.info(f'Stopping {host_role} in {HostInformation.get_os_name_and_version_from_inventory(inventory_path)}')
-        Executor.execute_command(inventory_path, f'systemctl stop {host_role}')
+        os_type = HostInformation.get_os_type(inventory_path)
+
+        if os_type == 'linux':
+            Executor.execute_commands(inventory_path, f'systemctl stop {host_role}')
+        elif os_type == 'windows':
+            Executor.execute_commands(inventory_path, f'NET STOP Wazuh')
 
 
     @staticmethod
@@ -562,7 +617,13 @@ class GeneralComponentActions:
         """
 
         logger.info(f'Restarting {host_role} in {HostInformation.get_os_name_and_version_from_inventory(inventory_path)}')
-        Executor.execute_command(inventory_path, f'systemctl restart {host_role}')
+        os_type = HostInformation.get_os_type(inventory_path)
+
+        if os_type == 'linux':
+            Executor.execute_commands(inventory_path, f'systemctl restart {host_role}')
+        elif os_type == 'windows':
+            Executor.execute_commands(inventory_path, 'NET STOP Wazuh')
+            Executor.execute_commands(inventory_path, 'NET START Wazuh')
 
 
     @staticmethod
@@ -576,7 +637,13 @@ class GeneralComponentActions:
         """
 
         logger.info(f'Starting {host_role} in {HostInformation.get_os_name_and_version_from_inventory(inventory_path)}')
-        Executor.execute_command(inventory_path, f'systemctl restart {host_role}')
+
+        os_type = HostInformation.get_os_type(inventory_path)
+
+        if os_type == 'linux':
+             Executor.execute_commands(inventory_path, f'systemctl start {host_role}')
+        elif os_type == 'windows':
+            Executor.execute_commands(inventory_path, 'NET START Wazuh')
 
 
     @staticmethod
@@ -590,7 +657,12 @@ class GeneralComponentActions:
         Returns:
             str: version
         """
-        return Executor.execute_command(inventory_path, f'{WAZUH_CONTROL} info -v')
+        os_type = HostInformation.get_os_type(inventory_path)
+
+        if os_type == 'linux':
+            return Executor.execute_commands(inventory_path, f'{WAZUH_CONTROL} info -v')
+        elif os_type == 'windows':
+            return Executor.execute_commands(inventory_path, f'Get-Content {WINDOWS_VERSION}'.replace("\n", ""))
 
 
     @staticmethod
@@ -604,7 +676,13 @@ class GeneralComponentActions:
         Returns:
             str: revision number
         """
-        return Executor.execute_command(inventory_path, f'{WAZUH_CONTROL} info -r')
+
+        os_type = HostInformation.get_os_type(inventory_path)
+
+        if os_type == 'linux':
+            return Executor.execute_commands(inventory_path, f'{WAZUH_CONTROL} info -r')
+        elif os_type == 'windows':
+            return Executor.execute_commands(inventory_path, f'Get-Content {WINDOWS_REVISION}'.replace("\n", ""))
 
 
     @staticmethod
@@ -618,7 +696,14 @@ class GeneralComponentActions:
         Returns:
             bool: True/False
         """
-        return 'true' in Executor.execute_command(inventory_path, f'[ -f {CLIENT_KEYS} ] && echo true || echo false')
+
+        os_type = HostInformation.get_os_type(inventory_path)
+
+        if os_type == 'linux':
+            return 'true' in Executor.execute_commands(inventory_path, f'[ -f {CLIENT_KEYS} ] && echo true || echo false')
+        elif os_type == 'windows':
+            return 'true' in Executor.execute_commands(inventory_path, f'Test-Path -Path "{CLIENT_KEYS}"')
+
 
 
     @staticmethod
@@ -633,7 +718,7 @@ class GeneralComponentActions:
         Returns:
             bool: True/False
         """
-        return 'active' == Executor.execute_command(inventory_path, f'systemctl is-active {host_role}').replace("\n", "")
+        return 'active' == Executor.execute_commands(inventory_path, f'systemctl is-active {host_role}').replace("\n", "")
 
 class Waits:
